requirements.txt
python-telegram-bot==20.6
beautifulsoup4
sumy
feedparser
deep-translator
apscheduler
nest_asyncio